requirements.txt
openai
httpx
langchain-openai
python-dotenv